+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2026-04-24
 - `14:10` -- **Updated** `libraries/ui/src/lib/supabaseApi.ts`
+- `14:10` -- **Updated** `libraries/ui/src/lib/supabaseApi.ts`
 - `14:09` -- **Updated** `libraries/ui/src/lib/supabaseApi.ts`
 - `14:06` -- **Updated** `apps/desktopProbe/src/pages/links.tsx`
 - `14:05` -- **Updated** `libraries/core/src/sdk.ts`